refactor: move HttpRequestMethod into types.ts
