+++ api/types.ts
@@ -1,1 +1,8 @@
 export type Path = { pathname: string }
+
+export enum HttpRequestMethod {
+  GET = 'GET',
+  POST = 'POST',
+  DELETE = 'DELETE',
+  PUT = 'PUT',
+}
